Fix xkcd comics made in January throwing an error
Previously, any commit written in January would throw an error because the array key for January was falsely '1' instead of the integer 1. This would lead to the moth name not being found and the error occuring.
This PR fixes that problem by changing the key of January to the integer 1.

**How to reproduce:**
Ask your CloudBot to `xkcd Regular Expressions`, an error will be shown in console.